create shared utils/helpers that others import
