+++ project/assets/textCleanup/utils/helpers.js
@@ -1,72 +1,6 @@
-import React from 'react';
-import _ from 'underscore';
-import moment from 'moment';
+import shared from 'shared/utils/helpers';
 
-
-var helpers = {
-    noop(){
-    },
-    fetchGet: {
-        credentials: 'same-origin',
-    },
-    fetchPost(csrf, obj, verb='POST'){
-        obj['csrfmiddlewaretoken'] = csrf;
-        return {
-            credentials: 'same-origin',
-            method: verb,
-            headers: new Headers({
-                'X-CSRFToken': csrf,
-                'content-type': 'application/json',
-            }),
-            body: JSON.stringify(obj),
-        };
-    },
-    fetchBulk(csrf, obj, verb='PATCH'){
-        obj['csrfmiddlewaretoken'] = csrf;
-        return {
-            credentials: 'same-origin',
-            method: verb,
-            headers: new Headers({
-                'X-CSRFToken': csrf,
-                'content-type': 'application/json',
-                'X-CUSTOM-BULK-OPERATION': true,
-            }),
-            body: JSON.stringify(obj),
-        };
-    },
-    fetchDelete(csrf){
-        return {
-            credentials: 'same-origin',
-            method: 'DELETE',
-            headers: new Headers({
-                'X-CSRFToken': csrf,
-                'content-type': 'application/json',
-            }),
-            body: JSON.stringify({csrfmiddlewaretoken:  csrf}),
-        };
-    },
-    getValue(target){
-        switch(target.type){
-        case 'checkbox':
-            return target.checked;
-        case 'number':
-            return parseFloat(target.value);
-        case 'select-one':  // use isFinite in-case value is 0
-            let val = parseInt(target.value);
-            return (_.isFinite(val)) ? val : target.value;
-        case 'text':
-        case 'textarea':
-        default:
-            return target.value;
-        }
-    },
-    getPatch(originalObj, newObj){
-        let patch = {};
-        _.each(newObj, function(v, k){
-            if (originalObj[k] !== v) patch[k] = v;
-        });
-        return patch;
-    },
+var helpers = Object.assign({}, shared, {
     getAssessmentApiUrl(config){
         return `${config.host}${config.assessment}?assessment_id=${config.assessment_id}`;
     },
@@ -76,30 +10,8 @@ var helpers = {
             idList = ids ? `&ids=${ids}` : '';
         return `${state.config.host}${state.config[state.router.params.type].url}${getFields}?assessment_id=${state.router.params.id}${fields}${idList}`;
     },
-    getObjectURL(base, id){
+    getObjectUrl(base, id){
         return `${base}${id}/`;
-    },
-    booleanCheckbox(val){
-        if (val){
-            return <i className='fa fa-check-square-o' title="checked"></i>;
-        } else {
-            return <i className='fa fa-square-o' title="un-checked"></i>;
-        }
-    },
-    datetimeFormat(dt){
-        return moment(dt).format('MMMM Do YYYY, h:mm:ss a');
-    },
-    goBack(e){
-        if (e && e.preventDefault) e.preventDefault();
-        window.history.back();
-    },
-    getInputDivClass(name, errors, extra=[]){
-        extra.push('form-group');
-        if (errors && errors[name]) extra.push('has-error');
-        return extra.join(' ');
-    },
-    deepCopy(object){
-        return JSON.parse(JSON.stringify(object));
     },
     caseToWords(string){
         return string
@@ -117,6 +29,6 @@ var helpers = {
     extendBreadcrumbs(url){
         $('.breadcrumb').children().last().contents().eq(-2).wrap(`<a href=${url}></a>`);
     },
-};
+});
 
 export default helpers;
